% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_zygous.R
\name{as_zygous}
\alias{as_zygous}
\title{Convert genotypes to zygosity}
\usage{
as_zygous(genotypes)
}
\arguments{
\item{genotypes}{A data table whose first column is the SNP identifier (a
character vector), and remaining columns refer to samples on which the SNP
has been genotyped. All columns except the first should be factors, and
should contain the levels \code{"AA"}, \code{"AB"}, \code{"BA"} and \code{"BB"}.}
}
\value{
A data table with the same structure as the data table passed in
\code{genotypes}. The only difference is that the sample columns (all except the
first) have their levels recoded to \code{"hom"} (homozygous) or \code{"het"}
(heterozygous).
}
\description{
This function takes a data table with genotypes (\code{"AA"}, \code{"AB"}, \code{"BA"} or
\code{"BB"}) and converts them to their corresponding zygosity, i.e., to
homozygous (\code{"hom"}) or heterozygous (\code{"het"}).

The argument \code{genotypes} is expected to be a data table whose first column is
the SNP identifier, and remaining columns refer to samples on which the SNP
has been genotyped. All columns except the first should be factors, and
should contain the levels \code{"AA"}, \code{"AB"}, \code{"BA"} and \code{"BB"}.

Note that this function changes its input by reference, meaning that the
variable passed as \code{genotypes} will be modified after running \code{as_zygous()}.
This is for performance reasons.
}
\examples{
levels <- c("AA", "AB", "BA", "BB")
(df <- data.frame(
  snp = c('rs123', 'rs456'),
  sample01 = factor(c('AB', 'BA'), levels),
  sample02 = factor(c('AA', 'BB'), levels)
))
as_zygous(df)
df

}
\seealso{
\code{\link[=read_snp_genotypes]{read_snp_genotypes()}}
}
