% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m2b.R
\name{b2m}
\alias{b2m}
\title{Convert Beta-values to M-values}
\usage{
b2m(b, base = 2)
}
\arguments{
\item{base}{The base with respect to which logarithms are computed. Defaults
to \code{base = 2}.}

\item{m}{A numeric vector of Beta-values.}
}
\description{
This function converts Beta-values to M-values.
}
\examples{
b2m(c(0, 0.5, 1))

}
\seealso{
\code{\link[=m2b]{m2b()}} for the reverse operation.
}
