% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daeqtl_mapping.R
\name{daeqtl_mapping}
\alias{daeqtl_mapping}
\title{DAEQTL mapping}
\usage{
daeqtl_mapping(
  snp_pairs,
  zygosity,
  ae,
  fn = daeqtl_test,
  ...,
  .extra_cols = 2L,
  .n_cores = 1L
)
}
\arguments{
\item{snp_pairs}{A data frame with DAE SNP/candidate SNP pairs.}

\item{zygosity}{A data frame with zygosity levels of samples. First column
should be the SNP identity, and remaining columns should refer to
biological samples. These should be factors, and their levels should be
\code{"hom"} (homozygous) and \code{"het"} (heterozygous).}

\item{ae}{A data frame with allelic expression (AE) ratios. Each row is for a
locus. The locus identity is indicated in the first column. Remaining
columns are samples.}

\item{fn}{A function implementing the statistical association approach.
This function needs to have two named arguments \code{ae_hom} and \code{ae_het}.
Extra arguments will be read from \code{...}.}

\item{...}{Extra arguments passed on to the call of \code{test_fn()}.}

\item{.extra_cols}{The number of extra columns you are creating with \code{fn}.}

\item{.n_cores}{The number of chunks to divide \code{snp_pairs} for parallel
processing.}
}
\value{
An updated version of the data frame \code{snp_pairs}. The update includes
extra columns, typically \code{pvalue} and \code{case}.
}
\description{
TODO
}
