% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daeqtl_plot.R
\name{daeqtl_plot}
\alias{daeqtl_plot}
\title{DAEQTL plot}
\usage{
daeqtl_plot(
  ae_hom,
  ae_het,
  dae_threshold = log2(1.5),
  min_n_hom = 2L,
  min_n_het = 2L,
  dae_threshold_linetype = "dashed",
  xlab = "Candidate SNP zygosity",
  ylab = "DAE SNP allelic expression (log-ratio)",
  ylim = NULL
)
}
\arguments{
\item{ae_hom}{Numeric vector of AE ratios of the DAE SNP. Each element of the
vector refers to a sample that is homozygous (\code{hom}) for the candidate
SNP.}

\item{ae_het}{Numeric vector of AE ratios of the DAE SNP. Each element of the
vector refers to a sample that is heterozygous (\code{het}) for the
candidate SNP.}

\item{dae_threshold}{An allelic expression (AE) threshold (in log-scale). A
sample showing an absolute AE greater than \code{dae_threshold} is considered
"technically" \emph{differential allelic expressed}, meaning that the
imbalance observed is not below the limit of detection. Adjustment made to
this parameter should depend on the experimental assay sensitivity.}

\item{min_n_hom}{Minimum number of samples in the homozygous group to be
considered eligible for statistical testing.}

\item{min_n_het}{Minimum number of samples in the heterozygous group to be
considered eligible for statistical testing.}

\item{dae_threshold_linetype}{Line type for the DAE threshold. See \href{https://ggplot2.tidyverse.org/articles/ggplot2-specs.html#lines}{Line type} for
options.}

\item{xlab}{Title of the x-axis.}

\item{ylab}{Title of the y-axis.}

\item{ylim}{A two element vector specifying the y-axis limits.}
}
\description{
A simple wrapper around a ggplot2 plot showing allelic expression ratios for
two groups of samples: homozygous and heterozygous. This is a handy function
to quickly inspect the data underlying each DAEQTL mapping test.
}
\examples{
ae_hom <- c(1.9, 2.1, 2 , 1.5, 1.4)
ae_het <- c(0.3, 0.6, 0.7)
daeqtl_plot(ae_hom = ae_hom, ae_het = ae_het)
}
