% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daeqtlr_example.R
\name{daeqtlr_example}
\alias{daeqtlr_example}
\title{Get path to daeqtlr example}
\usage{
daeqtlr_example(file = NULL)
}
\arguments{
\item{file}{Name of file. If \code{NULL}, the example files will be listed.}
}
\description{
daeqtlr comes bundled with a number of sample files in its \code{inst/extdata}
directory. This function make them easy to access.
}
\examples{
daeqtlr_example()
daeqtlr_example("zygosity.csv")
}
