% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_utils.R
\name{is_hom}
\alias{is_hom}
\alias{is_het}
\title{Who are the homozygous/heterozygous samples?}
\usage{
is_hom(zygosity, snp)

is_het(zygosity, snp)
}
\arguments{
\item{zygosity}{A data frame of zygosity levels: \code{"hom"} for homozygous or
\code{"het"} for heterozygous. Each row is for a locus. The locus identity is
indicated in the first column and named \code{snp}. Remaining columns are
samples.}

\item{snp}{String with SNP identifier.}
}
\value{
A logical vector.
}
\description{
This function determines who are the homozygous or heterozygous samples given
a data table of zygosity levels and a SNP of interest. Samples whose zygosity
level is \code{NA} will be returned as \code{FALSE}.
}
\examples{
# Let us start by reading in an example data set with zygosity levels
zygosity <- read_snp_zygosity(file = daeqtlr_example("zygosity.csv"))

# Checking out SNP rsX005
zygosity['rsX005']
is_hom(zygosity, 'rsX005')
is_het(zygosity, 'rsX005')

# Translate the logical vector to sample names
# Note that first column is excluded because it is the SNP identifier col
# Homozygous samples
(homs <- colnames(zygosity)[-1][is_hom(zygosity, 'rsX005')])

# Heterozygous samples
(hets <- colnames(zygosity)[-1][is_het(zygosity, 'rsX005')])

# Some samples are neither homozygous nor heterozygous because of NAs
# Note the `- 1` because the first column of `zygosity` is the SNP identifier.
(ncol(zygosity) - 1) - length(c(homs, hets))

# The samples whose zygosity is NA are:
setdiff(colnames(zygosity)[-1], c(homs, hets))

}
