% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m2b.R
\name{m2b}
\alias{m2b}
\title{Convert M-values to Beta-values}
\usage{
m2b(m, base = 2)
}
\arguments{
\item{m}{A numeric vector of M-values.}

\item{base}{The base with respect to which logarithms are computed. Defaults
to \code{base = 2}.}
}
\description{
This function converts M-values to Beta-values.
}
\examples{
m2b(c(-Inf, 0, Inf))

}
\seealso{
\code{\link[=b2m]{b2m()}} for the reverse operation.
}
