% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ae_ratios.R
\name{read_ae_ratios}
\alias{read_ae_ratios}
\title{Import expression ratios}
\usage{
read_ae_ratios(file, sep = ",", header = TRUE, ...)
}
\arguments{
\item{file}{A path to a file containing the expression ratios.}

\item{sep}{The separator between columns.}

\item{header}{Does the first data line contain column names?}

\item{...}{Extra arguments passed on to \code{\link[data.table:fread]{data.table::fread()}}.}
}
\value{
A data frame of expression values. Each row is for a locus. The locus
identity is indicated in the first column. Remaining columns are samples.
}
\description{
This function reads in expression ratios from a text file. The data is
expected to be in tabular format. First column is expected to be the locus
identity (e.g. SNP identifier), and remaining columns are the samples.
}
