% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_candidate_snps.R
\name{read_candidate_snps}
\alias{read_candidate_snps}
\title{Import DAEQTL candidates}
\usage{
read_candidate_snps(file, header = TRUE, distinct = TRUE)
}
\arguments{
\item{file}{A path to a file.}

\item{header}{Is the first line a line with a column name?}

\item{distinct}{Remove any duplicate SNPs?}
}
\value{
A one-column data frame. This single column is named \code{snp}.
  Rows are the candidate SNPs read from the text file.
}
\description{
This function reads in a list of candidate SNPs to being DAE quantitative
trait loci, from a text file. One SNP per row. By default, it expects a
header, which will be ignored. If the first line is immediately a SNP,
then use \code{header = FALSE} to not ignore it.
}
