% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dae_snps.R
\name{read_dae_snps}
\alias{read_dae_snps}
\title{Import DAE SNPs}
\usage{
read_dae_snps(file, header = TRUE, distinct = TRUE)
}
\arguments{
\item{file}{A path to a file.}

\item{header}{Is the first line a line with a column name?}

\item{distinct}{Remove any duplicate SNPs?}
}
\value{
A one-column data frame. This single column is named \code{dae_snp}.
  Rows are the DAE SNPs read from the text file.
}
\description{
This function reads in a list of DAE SNPs from a text file. One SNP per row.
By default, it expects a header, which will be ignored anyway. If the first
line is immediately a SNP, then use \code{header = FALSE}.
}
