% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_snp_gen_positions.R
\name{read_snp_gen_positions}
\alias{read_snp_gen_positions}
\title{Import SNP genomic positions}
\usage{
read_snp_gen_positions(file, sep = ",", header = TRUE, ...)
}
\arguments{
\item{file}{A path to a file containing the SNP genomic positions.}

\item{sep}{The separator between columns.}

\item{header}{Does the first data line contain column names?}

\item{...}{Extra arguments passed on to \code{\link[data.table:fread]{data.table::fread()}}.}
}
\value{
A data frame of genomic positions. Each row is for a locus. The locus
identity is indicated in the first column (\code{snp}). Second column is
the chromosome, and the third column is the position.
}
\description{
This function reads in genomic positions from a text file. The data is
expected to be in tabular format. First column is expected to be the locus
identity (e.g. SNP identifier), second column is the human chromosome (e.g.,
22 or X), and the third column is the position.
}
