% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_snp_genotypes.R
\name{read_snp_genotypes}
\alias{read_snp_genotypes}
\title{Import SNP genotypes}
\usage{
read_snp_genotypes(file, sep = ",", header = TRUE, ...)
}
\arguments{
\item{file}{A path to file.}

\item{sep}{The separator between columns.}

\item{header}{Does the first data line contain column names?}

\item{...}{Extra arguments to be passed on to \code{\link[data.table:fread]{data.table::fread()}}.}
}
\value{
A data frame of genotype values (\code{"AA"}, \code{"AB"}, \code{"BA"} or \code{"BB"}).
Each row is for a locus. The locus identity is indicated in the first
column and named \code{snp}. Remaining columns are samples.
}
\description{
This function reads in SNP genotypes from a text file. The data is expected
to be in tabular format. First column is expected to be the locus identity
(e.g. SNP identifier), and remaining columns are the samples.
}
