% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_snp_zygosity.R
\name{read_snp_zygosity}
\alias{read_snp_zygosity}
\title{Import SNP zygosity}
\usage{
read_snp_zygosity(file, sep = ",", header = TRUE, ...)
}
\arguments{
\item{file}{A path to file.}

\item{sep}{The separator between columns.}

\item{header}{Does the first data line contain column names?}

\item{...}{Extra arguments to be passed on to \code{\link[data.table:fread]{data.table::fread()}}.}
}
\value{
A data frame of zygosity levels: \code{"hom"} for homozygous or \code{"het"}
for heterozygous. Each row is for a locus. The locus identity is indicated
in the first column and named \code{snp}. Remaining columns are samples.
}
\description{
This function reads in SNP genotypes from a text file and converts them to
zygosity levels. The data is expected to be in tabular format. First column
is expected to be the locus identity (e.g. SNP identifier), and remaining
columns are the samples. The genotypes are expected to be encoded as \code{"AA"},
\code{"AB"}, \code{"BA"} or \code{"BB"}.
}
