% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_index.R
\name{split_index}
\alias{split_index}
\title{Create a split index}
\usage{
split_index(n, m)
}
\arguments{
\item{n}{Number of elements of the vector.}

\item{m}{Number of groups (or partitions).}
}
\value{
An integer vector.
}
\description{
This function creates an integer vector that can be used as a factor for
grouping with `split()`.
}
\keyword{internal}
