% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daeqtlr-package.R
\docType{package}
\name{daeqtlr-package}
\alias{daeqtlr}
\alias{daeqtlr-package}
\title{daeqtlr: DAEQTL Mapping}
\description{
Differential allelic expression (DAE) can be treated as a quantitative trait. Marker genotypes collected from the same set of individuals can, in turn, be used in testing statistical associations with DAE. This package provides routines for performing such a workflow, i.e. differential allelic expression trait loci mapping.
}
\author{
\strong{Maintainer}: Ramiro Magno \email{ramiro.magno@gmail.com} (\href{https://orcid.org/0000-0001-5226-3441}{ORCID})

Authors:
\itemize{
  \item Joana M Xavier \email{joana.xavier1@gmail.com} (\href{https://orcid.org/0000-0002-0702-6700}{ORCID})
  \item Ana-Teresa Maia \email{maia.anateresa@gmail.com} (\href{https://orcid.org/0000-0002-0454-9207}{ORCID})
}

Other contributors:
\itemize{
  \item CINTESIS [copyright holder, funder]
}

}
\keyword{internal}
